fix: Hyperlight pool config uses u64, add to README
- Fix type mismatch: guest_heap_size and guest_stack_size use u64
- Add Hyperlight to Platform Support section
- Add Hyperlight to Performance table (68ms latency)
- Add dedicated Hyperlight Backend section with requirements

+++ src/hyperlight_backend.rs
@@ -137,9 +137,9 @@ pub struct HyperlightPoolConfig {
     /// Maximum number of warm sandboxes to maintain
     pub max_warm: usize,
     /// Guest heap size in bytes
-    pub guest_heap_size: usize,
+    pub guest_heap_size: u64,
     /// Guest stack size in bytes
-    pub guest_stack_size: usize,
+    pub guest_stack_size: u64,
 }
 
 #[cfg(all(target_os = "linux", feature = "hyperlight"))]
@@ -343,8 +343,8 @@ pub struct HyperlightPoolStats {
 pub struct HyperlightPoolConfig {
     pub min_warm: usize,
     pub max_warm: usize,
-    pub guest_heap_size: usize,
-    pub guest_stack_size: usize,
+    pub guest_heap_size: u64,
+    pub guest_stack_size: u64,
 }
 
 /// Stub pool statistics
